Move `useTree` API to new package
