+++ apps/test-app/frontend/package.json
@@ -34,6 +34,7 @@
     "@itwin/presentation-components": "workspace:*",
     "@itwin/presentation-frontend": "^4.4.0",
     "@itwin/presentation-hierarchy-builder": "workspace:*",
+    "@itwin/presentation-hierarchies-react": "workspace:*",
     "@itwin/presentation-core-interop": "workspace:*",
     "@itwin/presentation-models-tree": "workspace:*",
     "@itwin/unified-selection": "workspace:*",
