+++ packages/hierarchies-react/.mocharc.json
@@ -1,0 +1,10 @@
+{
+  "require": ["ignore-styles"],
+  "checkLeaks": true,
+  "global": ["IS_REACT_ACT_ENVIRONMENT"],
+  "timeout": 60000,
+  "file": ["./lib/test/index.js"],
+  "exclude": ["./lib/test/coverage/**/*"],
+  "reporter": ["node_modules/@itwin/build-tools/mocha-reporter"],
+  "reporterOptions": ["mochaFile=lib/test/junit_results.xml"]
+}
