+++ packages/hierarchies-react/.npmignore
@@ -1,0 +1,14 @@
+# start off ignoring everything
+*
+# then add back only the files we want
+!*.md
+
+!lib/**/*.d.ts
+!lib/**/*.d.ts.map
+!lib/**/*.js
+!lib/**/*.js.map
+!lib/**/*.*css
+!lib/**/*.json
+
+# then ignore some stuff again
+lib/**/test/**
