+++ packages/hierarchies-react/package.json
@@ -1,0 +1,86 @@
+{
+  "name": "@itwin/presentation-hierarchies-react",
+  "version": "0.0.0",
+  "private": true,
+  "description": "React components based on `@itwin/presentation-hierarchies`",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/iTwin/presentation.git",
+    "directory": "packages/hierarchies-react"
+  },
+  "keywords": [
+    "Bentley",
+    "EC",
+    "Presentation",
+    "iTwin",
+    "iTwin.js",
+    "React",
+    "Component"
+  ],
+  "author": {
+    "name": "Bentley Systems, Inc.",
+    "url": "http://www.bentley.com"
+  },
+  "main": "lib/presentation-hierarchies-react.js",
+  "module": "lib/presentation-hierarchies-react.js",
+  "types": "lib/presentation-hierarchies-react.d.ts",
+  "scripts": {
+    "build": "npm run -s copy && npm run -s build:tsc",
+    "build:tsc": "tsc",
+    "build:watch": "npm run -s build:tsc -- -w",
+    "copy": "cpx \"./src/**/*.{css,json}\" ./lib",
+    "clean": "rimraf lib build",
+    "cover": "nyc npm -s test",
+    "lint": "eslint ./src/**/*.{ts,tsx}",
+    "test": "mocha --config ./.mocharc.json \"./lib/cjs/test/**/*.test.js\"",
+    "test:watch": "npm -s test -- --reporter min --watch-extensions ts,tsx --watch",
+    "extract-api": "betools extract-api --entry=presentation-hierarchies-react --apiReportFolder=./api --apiReportTempFolder=./api/temp --apiSummaryFolder=./api",
+    "check-internal": "node ../../scripts/checkInternal.js --apiSummary ./api/presentation-hierarchies-react.api.md"
+  },
+  "dependencies": {
+    "@itwin/itwinui-icons-react": "^2.8.0",
+    "@itwin/itwinui-illustrations-react": "^2.1.0",
+    "@itwin/itwinui-react": "^3.4.2",
+    "@itwin/presentation-hierarchy-builder": "workspace:*",
+    "classnames": "^2.3.2",
+    "immer": "^10.0.3",
+    "react-error-boundary": "^4.0.11",
+    "rxjs": "^7.8.1"
+  },
+  "peerDependencies": {
+    "@itwin/presentation-common": "^4.4.0",
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0"
+  },
+  "devDependencies": {
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/chai": "^4.3.11",
+    "@types/chai-as-promised": "^7.1.8",
+    "@types/chai-subset": "^1.3.5",
+    "@types/jsdom": "^21.1.6",
+    "@types/mocha": "^10.0.6",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
+    "@types/sinon": "^17.0.3",
+    "@types/sinon-chai": "^3.2.12",
+    "chai": "^4.4.1",
+    "chai-as-promised": "^7.1.1",
+    "chai-subset": "^1.6.0",
+    "cpx2": "^7.0.1",
+    "eslint": "^8.56.0",
+    "eslint-plugin-react": "^7.33.2",
+    "ignore-styles": "^5.0.1",
+    "jsdom": "^24.0.0",
+    "global-jsdom": "^24.0.0",
+    "mocha": "^10.3.0",
+    "nyc": "^15.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rimraf": "^5.0.5",
+    "sinon": "^17.0.1",
+    "sinon-chai": "^3.7.0",
+    "typescript": "~5.0.4"
+  }
+}
