+++ packages/hierarchies-react/src/presentation-hierarchies-react.ts
@@ -1,0 +1,9 @@
+export * from "./presentation-hierarchies-react/Tree";
+export * from "./presentation-hierarchies-react/Types";
+export * from "./presentation-hierarchies-react/UseTree";
+export {
+  UnifiedSelectionContextProvider,
+  EventLike,
+  UnifiedSelectionContainer,
+  UnifiedSelectionStore,
+} from "./presentation-hierarchies-react/UnifiedSelectionContext";
