+++ packages/hierarchies-react/src/presentation-hierarchies-react/internal/TreeModel.ts
