+++ packages/hierarchies-react/src/test/TestUtils.ts
@@ -1,0 +1,10 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+import { configure } from "@testing-library/react";
+
+configure({ reactStrictMode: true });
+
+export * from "@testing-library/react";
