+++ packages/hierarchies-react/tsconfig.json
@@ -1,6 +1,9 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
+    "jsx": "react-jsx",
+    "module": "ES2022",
+    "outDir": "lib",
     "sourceMap": true,
     "esModuleInterop": true,
     "inlineSources": true,
