+++ tsconfig.json
@@ -1,8 +1,8 @@
 {
   "compilerOptions": {
     "module": "CommonJS",
-    "target": "ES2020",
-    "lib": ["dom", "ES2020"],
+    "target": "ES2022",
+    "lib": ["dom", "ES2022"],
     "moduleResolution": "Node",
     "strict": true,
     "skipLibCheck": true,
